// collection-app/backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PackageStatus = {
  TO_BE_PICKED_UP: 'TO_BE_PICKED_UP',
  PICKED_UP: 'PICKED_UP',
  IN_TRANSIT: 'IN_TRANSIT',
  DELAYED: 'DELAYED',
  DELIVERED: 'DELIVERED'
} as const

export type PackageStatus = (typeof PackageStatus)[keyof typeof PackageStatus]